Smaller integer type is enough

+++ src/HDFWriterModule.cxx
@@ -22,7 +22,7 @@ HDFWriterModuleRegistry::find(HDFWriterModuleRegistry::K const &key) {
 
 namespace HDFWriterModule_detail {
 
-static std::map<uint64_t, std::string> const g_InitResult_strings{
+static std::map<int8_t, std::string> const g_InitResult_strings{
     {0, "OK"}, {-1, "ERROR_IO"}, {-2, "ERROR_INCOMPLETE_CONFIGURATION"},
 };
 
@@ -35,7 +35,7 @@ std::string InitResult::to_str() const {
   return it->second;
 }
 
-static std::map<uint64_t, std::string> const g_WriteResult_strings{
+static std::map<int8_t, std::string> const g_WriteResult_strings{
     {0, "OK"},
     {-1, "ERROR_IO"},
     {-2, "ERROR_BAD_FLATBUFFER"},
